requirements.txt
tflite-runtime
flask
numpy
pillow
gunicorn
